commit fix  6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,10 +10,10 @@ long int myLongInt;
 long long int myLInt;
 char myChar;
 float myFloat;
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(myChar));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(myInt));
-printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(myLongInt));
-printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(myLInt));
-printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(myFloat));
+printf("Size of a char: %lu byte(s)\n", sizeof(myChar));
+printf("Size of an int: %lu byte(s)\n", sizeof(myInt));
+printf("Size of long int: %lu byte(s)\n", sizeof(myLongInt));
+printf("Size of long long int: %lu byte(s)\n", sizeof(myLInt));
+printf("Size of float: %lu byte(s)\n", sizeof(myFloat));
 return (0);
 }
